refactor fswatch
