+++ .causality.json
@@ -1,0 +1,26 @@
+{
+  "desc": "Auto generated by fswatch [causal]",
+  "triggers": [
+    {
+      "name": "app-dev",
+      "pattens": [
+        "**/*.go",
+        "**/*.c",
+        "**/*.py"
+      ],
+      "env": {
+        "DEBUG": "1"
+      },
+      "cmd": "go test -v",
+      "shell": true,
+      "delay": "100ms",
+      "stop_timeout": "500ms",
+      "signal": "KILL",
+      "kill_signal": ""
+    }
+  ],
+  "watch_paths": [
+    "."
+  ],
+  "watch_depth": 0
+}
