+++ .causality.yml
@@ -1,0 +1,18 @@
+desc: Auto generated by fswatch [causality]
+triggers:
+- name: ""
+  pattens:
+  - '**/*.go'
+  - '**/*.c'
+  - '**/*.py'
+  env:
+    DEBUG: "1"
+  cmd: go test -v
+  shell: false
+  delay: 100ms
+  stop_timeout: 500ms
+  signal: KILL
+  kill_signal: ""
+watch_paths:
+- .
+watch_depth: 0
