--- .fsw.yml
@@ -1,17 +1,0 @@
-desc: Auto generated by fswatch [fswatch]
-triggers:
-- pattens:
-  - '*.go'
-  - '!testdata/*'
-  cmd: go test -v
-  name: test
-
-- pattens:
-  - 'testdata/*'
-  cmd: echo Haha, detect testdata changed && sleep 2
-  name: echo
-watch_paths:
-- .
-#watch_skips:
-#- '!vendor'
-watch_depth: 0
