+++ .gitignore
@@ -23,4 +23,4 @@ _cgo_export.*
 _testmain.go
 
 *.exe
-fswatch
+causality
