+++ .gopmfile
@@ -1,2 +1,2 @@
 [target]
-path=github.com/codeskyblue/fswatch
+path=github.com/sa-adebayo/causality
