+++ Godeps/Godeps.json
@@ -1,7 +1,7 @@
 {
-	"ImportPath": "github.com/codeskyblue/fswatch",
-	"GoVersion": "go1.6",
-	"GodepVersion": "v74",
+	"ImportPath": "github.com/sa-adebayo/causality",
+	"GoVersion": "go1.8",
+	"GodepVersion": "v79",
 	"Deps": [
 		{
 			"ImportPath": "github.com/codeskyblue/dockerignore",
