+++ appveyor.yml
@@ -3,7 +3,7 @@ version: "{build}"
 
 os: Windows Server 2012 R2
 
-clone_folder: c:\gopath\src\github.com\codeskyblue\fswatch
+clone_folder: c:\gopath\src\github.com\sa-adebayo\causality
 
 environment:
   GOPATH: c:\gopath
